refactor(conf): split config.go into focused files
Split the 2,802-line config.go into four files by responsibility:

- config.go (1,643 lines): type definitions, constants, pure methods
- migrations.go (470 lines): all Migrate* functions, model validation
- storage.go (562 lines): Load, Save, accessors, global state, viper init
- test_helpers.go (144 lines): SetTestSettings, GetTestSettings, SettingsBuilder

Zero behavior change. Same package, same exports, same function signatures.
Only file boundaries moved. All tests pass with race detector.
