+++ internal/conf/config.go
@@ -1,4 +1,4 @@
-// config.go: This file contains the configuration for the BirdNET-Go application. It defines the settings struct and functions to load and save the settings.
+// config.go defines the configuration types for the BirdNET-Go application.
 package conf
 
 import (
@@ -6,20 +6,11 @@ import (
 	"embed"
 	"encoding/base64"
 	"fmt"
-	"io/fs"
 	"iter"
-	"os"
-	"path/filepath"
-	"slices"
-	"strings"
-	"sync"
-	"sync/atomic"
 	"time"
 
-	"github.com/spf13/viper"
 	"github.com/tphakala/birdnet-go/internal/errors"
 	"github.com/tphakala/birdnet-go/internal/logger"
-	"gopkg.in/yaml.v3"
 )
 
 // Hemisphere detection thresholds
@@ -1568,841 +1559,6 @@ type AlertSettings struct {
 	HistoryRetentionDays int `json:"historyRetentionDays" yaml:"history_retention_days" mapstructure:"history_retention_days"` // Days to retain alert history (0 = unlimited)
 }
 
-// settingsInstance is the current settings instance
-// ConfigPath allows overriding the config file path via --config CLI flag.
-// Must be set before the first call to Setting() or Load().
-var ConfigPath string
-
-// settingsInstance holds the current *Settings snapshot. Published via
-// Store() in Load/StoreSettings/SetTestSettings; read via Load() in
-// GetSettings. Lock-free reads keep the request hot path fast while copy-on-
-// write writes in api/v2.UpdateSettings keep readers race-free.
-//
-// loadMu guards the on-demand Load() that Setting() performs when
-// settingsInstance is nil. A plain sync.Mutex is used rather than sync.Once
-// so that cleanup paths which call SetTestSettings(nil) can trigger a fresh
-// Load on the next Setting() call; sync.Once cannot be reset safely from a
-// parallel test without racing on the Once value itself.
-var (
-	settingsInstance atomic.Pointer[Settings]
-	loadMu           sync.Mutex
-)
-
-// persistMigration saves the config file after a successful migration.
-func persistMigration(settings *Settings, label string) {
-	configFile := viper.ConfigFileUsed()
-	if configFile == "" {
-		return
-	}
-	if err := SaveYAMLConfig(configFile, settings); err != nil {
-		GetLogger().Warn("Failed to save migrated "+label+" config", logger.Error(err))
-	} else {
-		GetLogger().Info("Saved migrated "+label+" configuration", logger.String("path", configFile))
-	}
-}
-
-// migrateStreamEnabledDefaults materializes missing enabled fields for legacy
-// RTSP stream configs before viper unmarshals them into the strongly typed
-// settings struct.
-func migrateStreamEnabledDefaults() bool {
-	normalizedStreams, migrated := normalizeRTSPStreamEnabledDefaults(viper.Get("realtime.rtsp.streams"))
-	if !migrated {
-		return false
-	}
-
-	viper.Set("realtime.rtsp.streams", normalizedStreams)
-	return true
-}
-
-// ensureSessionSecret backfills and persists security.sessionsecret for older
-// configs that predate the field so session signing remains stable across
-// restarts. Unlike the security package's temporary runtime fallback, this
-// updates the loaded settings and best-effort saves the generated secret.
-func ensureSessionSecret(settings *Settings) error {
-	if settings.Security.SessionSecret != "" {
-		return nil
-	}
-
-	sessionSecret, err := GenerateRandomSecret()
-	if err != nil {
-		return errors.New(err).
-			Component("conf").
-			Category(errors.CategoryConfiguration).
-			Context("operation", "generate_session_secret").
-			Build()
-	}
-
-	settings.Security.SessionSecret = sessionSecret
-
-	// Also set it in viper so it gets saved to config file
-	viper.Set("security.sessionsecret", sessionSecret)
-
-	// Log that we generated a new session secret
-	GetLogger().Info("Generated new SessionSecret for existing configuration")
-
-	// Save the updated config back to file to persist the generated secret
-	// This ensures the secret remains the same across restarts
-	configFile := viper.ConfigFileUsed()
-	if configFile == "" {
-		return nil
-	}
-
-	if err := SaveYAMLConfig(configFile, settings); err != nil {
-		// Log the error but don't fail - the generated secret will work for this session
-		GetLogger().Warn("Failed to save generated SessionSecret to config file", logger.Error(err))
-		return nil
-	}
-
-	// Set secure file permissions after saving
-	if err := os.Chmod(configFile, 0o600); err != nil {
-		GetLogger().Warn("Failed to set secure permissions on config file", logger.Error(err))
-	}
-
-	return nil
-}
-
-// Load reads the configuration file and environment variables into GlobalConfig.
-//
-//nolint:gocognit // Config loading is inherently complex; splitting adds indirection without clarity.
-func Load() (*Settings, error) {
-	// Create a new settings struct
-	settings := &Settings{}
-
-	// Initialize viper and read config
-	if err := initViper(); err != nil {
-		return nil, errors.New(err).
-			Category(errors.CategoryConfiguration).
-			Context("operation", "init-viper").
-			Build()
-	}
-
-	// Legacy stream entries omitted the enabled field entirely. Materialize the
-	// missing field before unmarshal so the runtime struct can use a plain bool.
-	streamEnabledMigrated := migrateStreamEnabledDefaults()
-
-	// Unmarshal the config into settings, with custom Duration decode hook
-	if err := viper.Unmarshal(settings, viper.DecodeHook(DurationDecodeHook())); err != nil {
-		return nil, errors.New(err).
-			Category(errors.CategoryConfiguration).
-			Context("operation", "unmarshal-config").
-			Build()
-	}
-
-	// Normalize species config keys to lowercase for case-insensitive matching
-	// This ensures that config keys like "American Robin" are converted to "american robin"
-	// to match the lowercase species names used in detection lookup (fixes #1701)
-	if settings.Realtime.Species.Config != nil {
-		settings.Realtime.Species.Config = NormalizeSpeciesConfigKeys(settings.Realtime.Species.Config)
-	}
-
-	// Migrate legacy AutoTLS boolean to new TLSMode field
-	if settings.MigrateTLSConfig() {
-		configFile := viper.ConfigFileUsed()
-		if configFile != "" {
-			if err := SaveYAMLConfig(configFile, settings); err != nil {
-				GetLogger().Warn("Failed to save migrated TLS config", logger.Error(err))
-			} else {
-				GetLogger().Info("Saved migrated TLS configuration", logger.String("path", configFile))
-			}
-		}
-	}
-
-	// Migrate legacy OAuth configuration to new array format
-	// This must happen before validation and saving
-	if settings.MigrateOAuthConfig() {
-		// Save the migrated config back to file
-		configFile := viper.ConfigFileUsed()
-		if configFile != "" {
-			if err := SaveYAMLConfig(configFile, settings); err != nil {
-				GetLogger().Warn("Failed to save migrated OAuth config", logger.Error(err))
-			} else {
-				GetLogger().Info("Saved migrated OAuth configuration", logger.String("path", configFile))
-			}
-		}
-	}
-
-	// Migrate legacy RTSP URLs to new streams format
-	if settings.MigrateRTSPConfig() {
-		configFile := viper.ConfigFileUsed()
-		if configFile != "" {
-			if err := SaveYAMLConfig(configFile, settings); err != nil {
-				GetLogger().Warn("Failed to save migrated RTSP config", logger.Error(err))
-			} else {
-				GetLogger().Info("Saved migrated RTSP configuration", logger.String("path", configFile))
-			}
-		}
-	}
-
-	// Migrate legacy single audio source to new multi-source format
-	if settings.MigrateAudioSourceConfig() {
-		persistMigration(settings, "audio source")
-	}
-
-	// Migrate per-source model field (model -> models)
-	if settings.MigrateSourceModels() {
-		persistMigration(settings, "source models")
-	}
-
-	if streamEnabledMigrated {
-		persistMigration(settings, "stream enabled defaults")
-	}
-
-	// Validate multi-model configuration
-	if err := settings.applyModelValidation(); err != nil {
-		return nil, err
-	}
-
-	// Migrate dashboard layout for existing installations
-	if settings.MigrateDashboardLayout() {
-		configFile := viper.ConfigFileUsed()
-		if configFile != "" {
-			if err := SaveYAMLConfig(configFile, settings); err != nil {
-				GetLogger().Warn("Failed to save migrated dashboard layout config", logger.Error(err))
-			} else {
-				GetLogger().Info("Saved migrated dashboard layout configuration", logger.String("path", configFile))
-			}
-		}
-	}
-
-	// Apply default transport to RTSP/RTMP streams that don't specify one
-	settings.Realtime.RTSP.ApplyStreamDefaults()
-
-	// Migrate LocationConfigured for backward compatibility with existing configs.
-	settings.MigrateLocationConfigured()
-
-	// Auto-generate SessionSecret if not set (for backward compatibility)
-	if err := ensureSessionSecret(settings); err != nil {
-		return nil, err
-	}
-
-	// Validate settings
-	if err := ValidateSettings(settings); err != nil {
-		// Check if it's just a validation warning (contains fallback info)
-		var validationErr ValidationError
-		if errors.As(err, &validationErr) {
-			// Report configuration issues to telemetry for debugging
-			for _, errMsg := range validationErr.Errors {
-				if strings.Contains(errMsg, "fallback") || strings.Contains(errMsg, "not supported") ||
-					strings.Contains(errMsg, "OAuth authentication warning") {
-					// This is a warning - report to telemetry but don't fail
-					GetLogger().Warn("Configuration warning", logger.String("message", errMsg))
-					// Store the warning for later telemetry reporting
-					settings.ValidationWarnings = append(settings.ValidationWarnings, errMsg)
-					// Note: Telemetry reporting will happen later in birdnet package when Sentry is initialized
-				} else {
-					// This is a real validation error - fail the config load
-					return nil, errors.New(err).
-						Category(errors.CategoryValidation).
-						Context("component", "settings").
-						Context("error_msg", errMsg).
-						Build()
-				}
-			}
-		} else {
-			// Other validation errors should fail the config load
-			return nil, errors.New(err).
-				Category(errors.CategoryValidation).
-				Context("component", "settings").
-				Build()
-		}
-	}
-
-	// Publish the loaded settings atomically. Readers calling GetSettings
-	// immediately after this point see this snapshot.
-	settingsInstance.Store(settings)
-	return settings, nil
-}
-
-// initViper initializes viper with default values and reads the configuration file.
-func initViper() error {
-	viper.SetConfigType("yaml")
-
-	// Configure environment variable support
-	if err := configureEnvironmentVariables(); err != nil {
-		// Log any validation warnings but don't fail startup
-		// This allows the application to continue with config file/default values
-		GetLogger().Warn("Environment variable configuration warning", logger.Error(err))
-	}
-
-	// Resolve effective config path. ConfigPath is the explicit override;
-	// fall back to scanning os.Args if it wasn't set yet (e.g., called
-	// before main parses args). Use a local variable to avoid mutating
-	// package-global state from the fallback parser.
-	effectiveConfigPath := ConfigPath
-	configFlagPresent := ConfigPath != ""
-	if !configFlagPresent {
-		for i, arg := range os.Args {
-			if (arg == "--config" || arg == "-c") && i+1 < len(os.Args) {
-				configFlagPresent = true
-				effectiveConfigPath = os.Args[i+1]
-				break
-			}
-			if val, found := strings.CutPrefix(arg, "--config="); found {
-				configFlagPresent = true
-				effectiveConfigPath = val
-				break
-			}
-			if val, found := strings.CutPrefix(arg, "-c="); found {
-				configFlagPresent = true
-				effectiveConfigPath = val
-				break
-			}
-		}
-	}
-
-	// Reject empty config path when the flag was explicitly provided
-	if configFlagPresent && effectiveConfigPath == "" {
-		return fmt.Errorf("--config flag requires a non-empty file path")
-	}
-
-	// If a custom config path was specified via --config, use it directly
-	if effectiveConfigPath != "" {
-		viper.SetConfigFile(effectiveConfigPath)
-	} else {
-		viper.SetConfigName("config")
-
-		// Get OS specific config paths
-		configPaths, err := GetDefaultConfigPaths()
-		if err != nil {
-			return errors.New(err).
-				Category(errors.CategoryConfiguration).
-				Context("operation", "get-config-paths").
-				Build()
-		}
-
-		// Assign config paths to Viper
-		for _, path := range configPaths {
-			viper.AddConfigPath(path)
-		}
-	}
-
-	// Set default values for each configuration parameter
-	// function defined in defaults.go
-	setDefaultConfig()
-
-	// Read configuration file
-	err := viper.ReadInConfig()
-	if err != nil {
-		// When an explicit config path was given, any read error is fatal —
-		// don't fall back to creating a default config elsewhere.
-		if effectiveConfigPath != "" {
-			return errors.New(err).
-				Category(errors.CategoryConfiguration).
-				Context("operation", "read-config-file").
-				Context("config_path", effectiveConfigPath).
-				Build()
-		}
-
-		// For default path search: ConfigFileNotFoundError means no config
-		// exists yet, so create one with defaults.
-		var configFileNotFoundError viper.ConfigFileNotFoundError
-		if errors.As(err, &configFileNotFoundError) {
-			return createDefaultConfig()
-		}
-
-		// Other errors (parse failures, permission issues) are fatal.
-		return errors.New(err).
-			Category(errors.CategoryFileIO).
-			Context("operation", "read-config-file").
-			Build()
-	}
-
-	return nil
-}
-
-// createDefaultConfig creates a default config file and writes it to the default config path
-func createDefaultConfig() error {
-	configPaths, err := GetDefaultConfigPaths()
-	if err != nil {
-		return errors.New(err).
-			Category(errors.CategoryConfiguration).
-			Context("operation", "create-default-config-paths").
-			Build()
-	}
-	configPath := filepath.Join(configPaths[0], "config.yaml")
-	defaultConfig := getDefaultConfig()
-
-	// If the basicauth secret is not set, generate a random one
-	if viper.GetString("security.basicauth.clientsecret") == "" {
-		clientSecret, err := GenerateRandomSecret()
-		if err != nil {
-			return errors.New(err).
-				Component("conf").
-				Category(errors.CategoryConfiguration).
-				Context("operation", "generate_client_secret").
-				Build()
-		}
-		viper.Set("security.basicauth.clientsecret", clientSecret)
-	}
-	// If the session secret is not set, generate a random one
-	// This ensures backward compatibility for existing deployments
-	if viper.GetString("security.sessionsecret") == "" {
-		sessionSecret, err := GenerateRandomSecret()
-		if err != nil {
-			return errors.New(err).
-				Component("conf").
-				Category(errors.CategoryConfiguration).
-				Context("operation", "generate_session_secret").
-				Build()
-		}
-		viper.Set("security.sessionsecret", sessionSecret)
-	}
-
-	// Create directories for config file
-	if err := os.MkdirAll(filepath.Dir(configPath), 0o750); err != nil {
-		return errors.New(err).
-			Category(errors.CategoryFileIO).
-			Context("operation", "create-config-dirs").
-			Context("path", filepath.Dir(configPath)).
-			Build()
-	}
-
-	// Write default config file with secure permissions (0600)
-	// Only the owner should be able to read/write the config file for security
-	if err := os.WriteFile(configPath, []byte(defaultConfig), 0o600); err != nil {
-		return errors.New(err).
-			Category(errors.CategoryFileIO).
-			Context("operation", "write-default-config").
-			Context("path", configPath).
-			Build()
-	}
-
-	fmt.Println("Created default config file at:", configPath)
-	return viper.ReadInConfig()
-}
-
-// getDefaultConfig reads the default configuration from the embedded config.yaml file.
-func getDefaultConfig() string {
-	data, err := fs.ReadFile(configFiles, "config.yaml")
-	if err != nil {
-		GetLogger().Error("Error reading config file", logger.Error(err))
-		os.Exit(1)
-	}
-	return string(data)
-}
-
-// GetSettings returns the current settings snapshot. Safe for concurrent use;
-// the returned pointer is published via atomic.Pointer.Store and never mutated
-// in place. Writers produce a new snapshot via CloneSettings + StoreSettings.
-func GetSettings() *Settings {
-	return settingsInstance.Load()
-}
-
-// CurrentOrFallback returns the latest published settings snapshot, or the
-// supplied fallback when none has been published. It exists so long-lived
-// services can pick up UI edits without a restart: capture the initial
-// *Settings in a field, then call conf.CurrentOrFallback(s.settings) whenever
-// the value is actually read. In production the fallback is effectively
-// unreachable (settings are always loaded before services start); it's there
-// so unit tests that inject a custom *Settings into a struct literal — without
-// touching the package-global atomic pointer — keep working.
-func CurrentOrFallback(fallback *Settings) *Settings {
-	if latest := settingsInstance.Load(); latest != nil {
-		return latest
-	}
-	return fallback
-}
-
-// StoreSettings publishes a new *Settings snapshot atomically. Callers must
-// not mutate the pointee after calling StoreSettings; readers may observe the
-// snapshot concurrently through GetSettings. The canonical writer pattern is:
-//
-//	current := conf.GetSettings()
-//	updated := conf.CloneSettings(current)
-//	// ... mutate updated ...
-//	conf.StoreSettings(updated)
-//
-// Passing nil is valid and clears the stored settings (mostly useful in tests).
-func StoreSettings(s *Settings) {
-	settingsInstance.Store(s)
-}
-
-// migrateLegacyProvider converts a legacy SocialProvider to the new OAuthProviderConfig format.
-// Returns nil if the legacy provider is not configured (no ClientID).
-func migrateLegacyProvider(providerName string, legacy SocialProvider) *OAuthProviderConfig {
-	if legacy.ClientID == "" {
-		return nil
-	}
-	return &OAuthProviderConfig{
-		Provider:     providerName,
-		Enabled:      legacy.Enabled,
-		ClientID:     legacy.ClientID,
-		ClientSecret: legacy.ClientSecret,
-		RedirectURI:  legacy.RedirectURI,
-		UserID:       legacy.UserId,
-	}
-}
-
-// MigrateTLSConfig migrates the legacy AutoTLS boolean to the new TLSMode field.
-// If TLSMode is already set, migration is skipped (already migrated or user has
-// explicitly configured the new field). Returns true if migration occurred.
-func (s *Settings) MigrateTLSConfig() bool {
-	// Skip if TLSMode is already set (already migrated or explicitly configured)
-	if s.Security.TLSMode != TLSModeNone {
-		return false
-	}
-
-	// Migrate AutoTLS=true to TLSMode="autotls"
-	if s.Security.AutoTLS {
-		s.Security.TLSMode = TLSModeAutoTLS
-		return true
-	}
-
-	return false
-}
-
-// MigrateOAuthConfig migrates legacy OAuth configuration (GoogleAuth, GithubAuth, MicrosoftAuth)
-// to the new OAuthProviders array format. This migration:
-// - Skips if OAuthProviders already has entries (already migrated)
-// - Only migrates providers that have a ClientID configured
-// - Preserves all settings from the legacy format
-// - Returns true if migration occurred, false if skipped
-func (s *Settings) MigrateOAuthConfig() bool {
-	// Skip if already migrated (new array has entries)
-	if len(s.Security.OAuthProviders) > 0 {
-		return false
-	}
-
-	// Define legacy providers to migrate
-	legacyProviders := []struct {
-		name   string
-		config SocialProvider
-	}{
-		{"google", s.Security.GoogleAuth},
-		{"github", s.Security.GithubAuth},
-		{"microsoft", s.Security.MicrosoftAuth},
-	}
-
-	var migrated bool
-	for _, legacy := range legacyProviders {
-		if cfg := migrateLegacyProvider(legacy.name, legacy.config); cfg != nil {
-			s.Security.OAuthProviders = append(s.Security.OAuthProviders, *cfg)
-			migrated = true
-			GetLogger().Info("Migrated OAuth configuration to new format", logger.String("provider", legacy.name))
-		}
-	}
-
-	if migrated {
-		GetLogger().Info("OAuth configuration migration complete", logger.String("note", "Legacy fields will be ignored"))
-	}
-
-	return migrated
-}
-
-// inferStreamType detects the stream type from URL scheme.
-// Returns StreamTypeRTSP as default for unknown schemes.
-func inferStreamType(url string) string {
-	urlLower := strings.ToLower(url)
-
-	switch {
-	case strings.HasPrefix(urlLower, "rtsp://"), strings.HasPrefix(urlLower, "rtsps://"):
-		return StreamTypeRTSP
-	case strings.HasPrefix(urlLower, "rtmp://"), strings.HasPrefix(urlLower, "rtmps://"):
-		return StreamTypeRTMP
-	case strings.HasPrefix(urlLower, "udp://"), strings.HasPrefix(urlLower, "rtp://"):
-		return StreamTypeUDP
-	case strings.HasPrefix(urlLower, "http://"), strings.HasPrefix(urlLower, "https://"):
-		// Check for HLS (.m3u8) vs generic HTTP
-		if strings.Contains(urlLower, ".m3u8") {
-			return StreamTypeHLS
-		}
-		return StreamTypeHTTP
-	default:
-		return StreamTypeRTSP // Default to RTSP for unknown schemes
-	}
-}
-
-// MigrateRTSPConfig migrates legacy URLs []string to Streams []StreamConfig.
-// This migration:
-// - Skips if Streams already has entries (already migrated)
-// - Only migrates if URLs has data
-// - Trims whitespace and skips empty URLs
-// - Infers stream type from URL scheme
-// - Preserves the global Transport setting for RTSP/RTMP streams
-// - Returns true if migration occurred, false if skipped
-func (s *Settings) MigrateRTSPConfig() bool {
-	rtsp := &s.Realtime.RTSP
-
-	// Skip if already migrated (new format has streams)
-	if len(rtsp.Streams) > 0 {
-		return false
-	}
-
-	// Skip if no legacy URLs to migrate
-	if len(rtsp.URLs) == 0 {
-		return false
-	}
-
-	// Get global transport, default to tcp
-	globalTransport := rtsp.Transport
-	if globalTransport == "" {
-		globalTransport = DefaultTransport
-	}
-
-	// Preallocate streams slice with capacity and track seen URLs for deduplication
-	rtsp.Streams = make([]StreamConfig, 0, len(rtsp.URLs))
-	seenURLs := make(map[string]bool)
-	streamIndex := 0
-
-	// Migrate each URL to StreamConfig
-	for _, rawURL := range rtsp.URLs {
-		// Trim whitespace and skip empty URLs
-		url := strings.TrimSpace(rawURL)
-		if url == "" {
-			continue
-		}
-
-		// Skip duplicate URLs to ensure valid configuration
-		if seenURLs[url] {
-			continue
-		}
-		seenURLs[url] = true
-		streamIndex++
-
-		// Infer stream type from URL scheme
-		streamType := inferStreamType(url)
-
-		// Only apply transport setting to RTSP/RTMP types where it makes sense
-		transport := ""
-		if streamType == StreamTypeRTSP || streamType == StreamTypeRTMP {
-			transport = globalTransport
-		}
-
-		stream := StreamConfig{
-			Name:      fmt.Sprintf("Stream %d", streamIndex),
-			URL:       url,
-			Enabled:   true,
-			Type:      streamType,
-			Transport: transport,
-		}
-		rtsp.Streams = append(rtsp.Streams, stream)
-	}
-
-	// If no valid URLs were found, don't mark as migrated
-	if len(rtsp.Streams) == 0 {
-		return false
-	}
-
-	// Clear legacy fields
-	rtsp.URLs = nil
-	rtsp.Transport = ""
-
-	GetLogger().Info("Migrated RTSP configuration to new streams format",
-		logger.Int("stream_count", len(rtsp.Streams)))
-
-	return true
-}
-
-// MigrateAudioSourceConfig migrates the legacy single audio source string
-// (realtime.audio.source) to the new multi-source array format (realtime.audio.sources).
-// This follows the same pattern as MigrateRTSPConfig.
-func (s *Settings) MigrateAudioSourceConfig() bool {
-	audio := &s.Realtime.Audio
-
-	// Skip if already migrated (new format has sources) OR nothing to migrate.
-	// Note: Viper unmarshals explicit `sources: []` as empty slice (not nil),
-	// so we check both conditions to avoid re-migrating.
-	if len(audio.Sources) > 0 || strings.TrimSpace(audio.Source) == "" {
-		return false
-	}
-
-	// Create one AudioSourceConfig from the legacy scalar
-	audio.Sources = []AudioSourceConfig{{
-		Name:       "Sound Card 1",
-		Device:     strings.TrimSpace(audio.Source),
-		Gain:       0,
-		QuietHours: audio.QuietHours, // Carry over global quiet hours
-	}}
-
-	// Clear legacy field
-	audio.Source = ""
-
-	GetLogger().Info("Migrated audio source to new multi-source format",
-		logger.Int("source_count", 1))
-
-	return true
-}
-
-// MigrateSourceModels migrates the legacy singular Model field to the new
-// Models list on AudioSourceConfig and StreamConfig. Sources with neither
-// Model nor Models set default to ["birdnet"]. Returns true if any migration
-// occurred.
-func (s *Settings) MigrateSourceModels() bool {
-	migrated := false
-
-	for i := range s.Realtime.Audio.Sources {
-		src := &s.Realtime.Audio.Sources[i]
-		if len(src.Models) > 0 {
-			continue
-		}
-		if src.Model != "" {
-			src.Models = []string{src.Model}
-			src.Model = ""
-		} else {
-			src.Models = []string{"birdnet"}
-		}
-		migrated = true
-	}
-
-	for _, stream := range s.Realtime.RTSP.AllStreams() {
-		if len(stream.Models) > 0 {
-			continue
-		}
-		stream.Models = []string{"birdnet"}
-		migrated = true
-	}
-
-	return migrated
-}
-
-// normalizeRTSPStreamEnabledDefaults materializes enabled=true for legacy raw
-// RTSP stream entries that predate the field. It operates on Viper's raw data
-// before unmarshal so StreamConfig can keep a plain bool.
-func normalizeRTSPStreamEnabledDefaults(rawStreams any) ([]any, bool) {
-	streams, ok := rawStreams.([]any)
-	if !ok {
-		return nil, false
-	}
-
-	normalized := make([]any, len(streams))
-	migrated := false
-
-	for i, rawStream := range streams {
-		streamMap, ok := rawStream.(map[string]any)
-		if !ok {
-			normalized[i] = rawStream
-			continue
-		}
-		if val, exists := streamMap["enabled"]; exists && val != nil {
-			normalized[i] = rawStream
-			continue
-		}
-
-		copied := make(map[string]any, len(streamMap)+1)
-		for key, value := range streamMap {
-			copied[key] = value
-		}
-		copied["enabled"] = true
-		normalized[i] = copied
-		migrated = true
-	}
-
-	if !migrated {
-		return nil, false
-	}
-
-	return normalized, true
-}
-
-// ValidateModelConfig checks model-related configuration for errors and
-// warnings. Returns a slice of warning/error strings. Fatal errors are
-// prefixed with "error:"; non-fatal issues with "warning:".
-// knownIDs is the set of recognized config-level model identifiers; callers
-// should pass classifier.KnownConfigIDs() when available, or a hardcoded
-// fallback during early config loading.
-func (s *Settings) ValidateModelConfig(knownIDs map[string]bool) []string {
-	var issues []string
-
-	enabledSet := make(map[string]bool, len(s.Models.Enabled))
-	for _, id := range s.Models.Enabled {
-		enabledSet[strings.ToLower(id)] = true
-	}
-
-	for _, id := range s.Models.Enabled {
-		if !knownIDs[strings.ToLower(id)] {
-			issues = append(issues, "warning: unknown model ID in models.enabled: "+id)
-		}
-	}
-
-	if enabledSet["perch_v2"] && !s.Perch.Enabled {
-		issues = append(issues, "error: perch_v2 in models.enabled but perch.enabled is false")
-	}
-
-	if s.Perch.Enabled && !enabledSet["perch_v2"] {
-		issues = append(issues, "warning: perch.enabled is true but 'perch_v2' is not in models.enabled")
-	}
-
-	if s.Perch.Enabled {
-		if s.Perch.ModelPath == "" {
-			issues = append(issues, "error: perch.enabled is true but perch.modelpath is empty")
-		}
-		if s.Perch.LabelPath == "" {
-			issues = append(issues, "error: perch.enabled is true but perch.labelpath is empty")
-		}
-	}
-
-	for i := range s.Realtime.Audio.Sources {
-		src := &s.Realtime.Audio.Sources[i]
-		for _, modelID := range src.Models {
-			if !enabledSet[strings.ToLower(modelID)] {
-				issues = append(issues, "warning: source \""+src.Name+"\" references model \""+modelID+"\" not in models.enabled")
-			}
-		}
-	}
-	for _, stream := range s.Realtime.RTSP.AllStreams() {
-		for _, modelID := range stream.Models {
-			if !enabledSet[strings.ToLower(modelID)] {
-				issues = append(issues, "warning: stream \""+stream.Name+"\" references model \""+modelID+"\" not in models.enabled")
-			}
-		}
-	}
-
-	return issues
-}
-
-// applyModelValidation runs ValidateModelConfig and either returns an error
-// for fatal issues or appends warnings to ValidationWarnings. All fatal
-// errors are collected and returned together so the user can fix them in
-// one pass.
-func (s *Settings) applyModelValidation() error {
-	// Default known IDs — matches classifier.KnownConfigIDs() at compile time.
-	// This fallback is used during config loading before the classifier package
-	// is available. The orchestrator re-validates with the authoritative list.
-	knownIDs := map[string]bool{"birdnet": true, "perch_v2": true}
-	modelIssues := s.ValidateModelConfig(knownIDs)
-	var fatalErrors []string
-	for _, issue := range modelIssues {
-		if strings.HasPrefix(issue, "error:") {
-			fatalErrors = append(fatalErrors, strings.TrimPrefix(issue, "error: "))
-		} else {
-			GetLogger().Warn("model configuration issue", logger.String("issue", issue))
-			s.ValidationWarnings = append(s.ValidationWarnings, issue)
-		}
-	}
-	if len(fatalErrors) > 0 {
-		return errors.Newf("model configuration: %s", strings.Join(fatalErrors, "; ")).
-			Category(errors.CategoryValidation).
-			Build()
-	}
-	return nil
-}
-
-// MigrateLocationConfigured sets LocationConfigured to true for existing configs
-// that have non-zero coordinates but predate the explicit flag. This provides
-// backward compatibility so that existing users don't lose location-dependent
-// features after upgrading. New installations start with LocationConfigured=false
-// until the user explicitly sets coordinates (even 0,0).
-func (s *Settings) MigrateLocationConfigured() {
-	if s.BirdNET.LocationConfigured {
-		return
-	}
-
-	if s.BirdNET.Latitude == 0 && s.BirdNET.Longitude == 0 {
-		return
-	}
-
-	s.BirdNET.LocationConfigured = true
-
-	configFile := viper.ConfigFileUsed()
-	if configFile != "" {
-		if err := SaveYAMLConfig(configFile, s); err != nil {
-			GetLogger().Warn("Failed to save migrated LocationConfigured flag", logger.Error(err))
-		} else {
-			GetLogger().Info("Migrated LocationConfigured flag based on existing coordinates")
-		}
-	}
-}
-
 // GetOAuthProvider returns the OAuth provider configuration for the given provider ID.
 // Returns nil if the provider is not configured.
 func (s *Settings) GetOAuthProvider(providerID string) *OAuthProviderConfig {
@@ -2430,322 +1586,6 @@ func (s *Settings) GetEnabledOAuthProviders() []string {
 		}
 	}
 	return enabled
-}
-
-// prepareSettingsForSave applies data transformations to settings before saving.
-// This function is separated from SaveSettings to enable unit testing without filesystem I/O.
-//
-// Current transformations:
-//   - Auto-populates seasonal tracking seasons based on latitude if not already set
-//
-// Note: This is a pure function that only transforms data. It does not handle:
-//   - Mutex locking (handled by SaveSettings caller)
-//   - File I/O operations (handled by SaveSettings)
-//   - Species list synchronization (handled separately in SaveSettings)
-func prepareSettingsForSave(s *Settings, latitude float64) Settings {
-	settingsCopy := *s
-
-	// Auto-update seasonal tracking dates based on latitude if seasonal tracking is enabled
-	// and no custom seasons are already defined
-	if settingsCopy.Realtime.SpeciesTracking.SeasonalTracking.Enabled &&
-		len(settingsCopy.Realtime.SpeciesTracking.SeasonalTracking.Seasons) == 0 {
-		// Get hemisphere-appropriate default seasons
-		defaultSeasons := GetDefaultSeasons(latitude)
-		settingsCopy.Realtime.SpeciesTracking.SeasonalTracking.Seasons = defaultSeasons
-	}
-
-	return settingsCopy
-}
-
-// SaveSettings saves the current settings to the configuration file.
-// It uses UpdateYAMLConfig to handle the atomic write process.
-func SaveSettings() error {
-	// Load the current snapshot through the atomic pointer; the pointee is
-	// immutable after publication so reading fields off it is race-free.
-	current := settingsInstance.Load()
-	if current == nil {
-		return errors.Newf("settings not initialized").
-			Category(errors.CategoryConfiguration).
-			Context("operation", "save-settings").
-			Build()
-	}
-
-	// Deep-clone the published snapshot before mutating it in
-	// prepareSettingsForSave. A shallow copy would leave slice and map
-	// backing arrays shared with concurrent readers that hold current;
-	// any transformation that mutates one of those nested fields would
-	// race against them.
-	settingsCopy := *CloneSettings(current)
-
-	// Refresh the species list under its dedicated lock so the runtime
-	// list (shared across Species.Include/Exclude updates) is captured
-	// atomically rather than via the clone that predates the save.
-	speciesListMutex.RLock()
-	settingsCopy.BirdNET.RangeFilter.Species = slices.Clone(current.BirdNET.RangeFilter.Species)
-	speciesListMutex.RUnlock()
-
-	// Apply data transformations (seasonal tracking, etc.) on the clone.
-	settingsCopy = prepareSettingsForSave(&settingsCopy, current.BirdNET.Latitude)
-
-	// Find the path of the current config file
-	configPath, err := FindConfigFile()
-	if err != nil {
-		return errors.New(err).
-			Category(errors.CategoryFileIO).
-			Context("operation", "find-config-file").
-			Build()
-	}
-
-	// Save the settings to the config file
-	if err := SaveYAMLConfig(configPath, &settingsCopy); err != nil {
-		return errors.New(err).
-			Category(errors.CategoryFileIO).
-			Context("operation", "save-yaml-config").
-			Context("path", configPath).
-			Build()
-	}
-
-	GetLogger().Info("Settings saved successfully", logger.String("path", configPath))
-	return nil
-}
-
-// Setting returns the current settings instance, initializing it if necessary
-func Setting() *Settings {
-	// Fast path: settings already published.
-	if s := settingsInstance.Load(); s != nil {
-		return s
-	}
-	// Slow path: lazy-load from disk. Serialise concurrent first-callers
-	// through loadMu; the atomic check inside the lock collapses the extras.
-	loadMu.Lock()
-	if s := settingsInstance.Load(); s != nil {
-		loadMu.Unlock()
-		return s
-	}
-	if _, err := Load(); err != nil {
-		// Fatal error loading settings - application cannot continue.
-		// Release the lock explicitly because os.Exit does not run defers.
-		enhancedErr := errors.New(err).
-			Category(errors.CategoryConfiguration).
-			Context("operation", "load-settings-init").
-			Build()
-		GetLogger().Error("Error loading settings", logger.Error(enhancedErr))
-		loadMu.Unlock()
-		os.Exit(1)
-	}
-	loadMu.Unlock()
-	return settingsInstance.Load()
-}
-
-// SetTestSettings allows tests to inject their own settings instance.
-// This must be called before any call to Setting() to be effective.
-// This is intended for testing purposes only.
-func SetTestSettings(settings *Settings) {
-	// Publish atomically; readers on the hot path see the new snapshot
-	// immediately via GetSettings. Setting() will re-Load() from disk on
-	// the next call if settings is nil, so cleanup paths that restore a
-	// previously-nil snapshot do not leave downstream consumers with a
-	// nil dereference.
-	settingsInstance.Store(settings)
-}
-
-// GetTestSettings returns a copy of default settings suitable for testing.
-// This creates isolated settings that won't affect the global configuration.
-func GetTestSettings() *Settings {
-	settings := &Settings{}
-
-	// Initialize with defaults
-	settings.Debug = false
-	settings.Main.Name = "BirdNET-Go-Test"
-	settings.Main.TimeAs24h = true
-
-	// Set up minimal test configuration
-	settings.BirdNET.Sensitivity = 1.0
-	settings.BirdNET.Threshold = 0.8
-	settings.BirdNET.Overlap = 0.0
-	settings.BirdNET.Locale = "en"
-
-	// Dashboard settings with thumbnails
-	settings.Realtime.Dashboard.Thumbnails.Debug = false
-	settings.Realtime.Dashboard.Thumbnails.Summary = false
-	settings.Realtime.Dashboard.Thumbnails.Recent = true
-	settings.Realtime.Dashboard.Thumbnails.ImageProvider = "avicommons"
-	settings.Realtime.Dashboard.Thumbnails.FallbackPolicy = "none"
-
-	// Other realtime settings
-	settings.Realtime.Interval = 15
-	settings.Realtime.ProcessingTime = false
-
-	// Web server settings
-	settings.WebServer.Enabled = false
-	settings.WebServer.Port = "8080"
-
-	// Output settings
-	settings.Output.SQLite.Enabled = false
-	settings.Output.SQLite.Path = ":memory:"
-
-	return settings
-}
-
-// SettingsBuilder provides a fluent interface for constructing test settings.
-// It simplifies test setup by providing convenient methods for common configuration patterns.
-//
-// Example usage:
-//
-//	settings := conf.NewTestSettings().
-//	    WithBirdNET(0.9, 45.0, -122.0).
-//	    WithMQTT("tcp://localhost:1883", "test").
-//	    Build()
-type SettingsBuilder struct {
-	settings *Settings
-}
-
-// NewTestSettings creates a new SettingsBuilder initialized with default test settings.
-func NewTestSettings() *SettingsBuilder {
-	return &SettingsBuilder{
-		settings: GetTestSettings(),
-	}
-}
-
-// WithBirdNET configures BirdNET-specific settings.
-func (b *SettingsBuilder) WithBirdNET(threshold, latitude, longitude float64) *SettingsBuilder {
-	b.settings.BirdNET.Threshold = threshold
-	b.settings.BirdNET.Latitude = latitude
-	b.settings.BirdNET.Longitude = longitude
-	return b
-}
-
-// WithMQTT configures MQTT settings and enables MQTT.
-func (b *SettingsBuilder) WithMQTT(broker, topic string) *SettingsBuilder {
-	b.settings.Realtime.MQTT.Enabled = true
-	b.settings.Realtime.MQTT.Broker = broker
-	b.settings.Realtime.MQTT.Topic = topic
-	return b
-}
-
-// WithAudioExport configures audio export settings and enables audio export.
-func (b *SettingsBuilder) WithAudioExport(path, exportType, bitrate string) *SettingsBuilder {
-	b.settings.Realtime.Audio.Export.Enabled = true
-	b.settings.Realtime.Audio.Export.Path = path
-	b.settings.Realtime.Audio.Export.Type = exportType
-	b.settings.Realtime.Audio.Export.Bitrate = bitrate
-	return b
-}
-
-// WithSpeciesTracking configures species tracking settings and enables species tracking.
-func (b *SettingsBuilder) WithSpeciesTracking(windowDays, syncInterval int) *SettingsBuilder {
-	b.settings.Realtime.SpeciesTracking.Enabled = true
-	b.settings.Realtime.SpeciesTracking.NewSpeciesWindowDays = windowDays
-	b.settings.Realtime.SpeciesTracking.SyncIntervalMinutes = syncInterval
-	return b
-}
-
-// WithRTSPHealthThreshold configures RTSP health monitoring threshold.
-func (b *SettingsBuilder) WithRTSPHealthThreshold(seconds int) *SettingsBuilder {
-	b.settings.Realtime.RTSP.Health.HealthyDataThreshold = seconds
-	return b
-}
-
-// WithImageProvider configures thumbnail image provider settings.
-func (b *SettingsBuilder) WithImageProvider(provider, fallbackPolicy string) *SettingsBuilder {
-	b.settings.Realtime.Dashboard.Thumbnails.ImageProvider = provider
-	b.settings.Realtime.Dashboard.Thumbnails.FallbackPolicy = fallbackPolicy
-	return b
-}
-
-// WithSecurity configures security settings.
-func (b *SettingsBuilder) WithSecurity(host string, autoTLS bool) *SettingsBuilder {
-	b.settings.Security.Host = host
-	b.settings.Security.AutoTLS = autoTLS
-	return b
-}
-
-// WithWebServer configures web server settings.
-func (b *SettingsBuilder) WithWebServer(port string, enabled bool) *SettingsBuilder {
-	b.settings.WebServer.Port = port
-	b.settings.WebServer.Enabled = enabled
-	return b
-}
-
-// Build returns the constructed settings without modifying global state.
-// Use this when you need the settings object for manual manipulation.
-func (b *SettingsBuilder) Build() *Settings {
-	return b.settings
-}
-
-// Apply sets the built settings as the global test settings.
-// This is equivalent to calling SetTestSettings() with the built settings.
-func (b *SettingsBuilder) Apply() *Settings {
-	SetTestSettings(b.settings)
-	return b.settings
-}
-
-// Note: SendValidationWarningsAsNotifications function removed as it was unused
-
-// SaveYAMLConfig updates the YAML configuration file with new settings.
-// It overwrites the existing file, not preserving comments or structure.
-func SaveYAMLConfig(configPath string, settings *Settings) error {
-	// Marshal the settings struct to YAML
-	yamlData, err := yaml.Marshal(settings)
-	if err != nil {
-		return errors.New(err).
-			Category(errors.CategoryConfiguration).
-			Context("operation", "yaml-marshal").
-			Build()
-	}
-
-	// Write the YAML data to a temporary file
-	// This is done to ensure atomic write operation
-	tempFile, err := os.CreateTemp(filepath.Dir(configPath), "config-*.yaml")
-	if err != nil {
-		return errors.New(err).
-			Category(errors.CategoryFileIO).
-			Context("operation", "create-temp-file").
-			Context("dir", filepath.Dir(configPath)).
-			Build()
-	}
-	tempFileName := tempFile.Name()
-	// Ensure the temporary file is removed in case of any failure
-	defer func() {
-		if err := os.Remove(tempFileName); err != nil && !os.IsNotExist(err) {
-			GetLogger().Warn("Failed to remove temporary file", logger.Error(err), logger.String("file", tempFileName))
-		}
-	}()
-
-	// Write the YAML data to the temporary file
-	if _, err := tempFile.Write(yamlData); err != nil {
-		// Best effort close on error path
-		_ = tempFile.Close()
-		return errors.New(err).
-			Category(errors.CategoryFileIO).
-			Context("operation", "write-temp-file").
-			Build()
-	}
-	// Close the temporary file after writing
-	if err := tempFile.Close(); err != nil {
-		return errors.New(err).
-			Category(errors.CategoryFileIO).
-			Context("operation", "close-temp-file").
-			Build()
-	}
-
-	// Try to rename the temporary file to replace the original config file
-	// This is typically an atomic operation on most filesystems
-	if err := os.Rename(tempFileName, configPath); err != nil {
-		// If rename fails (e.g., cross-device link), fall back to copy & delete
-		// This might happen when the temp directory is on a different filesystem
-		if err := moveFile(tempFileName, configPath); err != nil {
-			return errors.New(err).
-				Category(errors.CategoryFileIO).
-				Context("operation", "move-config-file").
-				Context("src", tempFileName).
-				Context("dst", configPath).
-				Build()
-		}
-	}
-
-	// If we've reached this point, the operation was successful
-	return nil
 }
 
 // GenerateRandomSecret generates a URL-safe base64 encoded random string
